# Changes to mu_ij if statements for performance improvements