Changes made to the force calculation for performance improvement. 

Results below based on the KalthoffWinkler example with: 
- std::array<int, 3> num_cell = { 400, 800, 36 };
- double delta = 0.00150000001;
running on 1 node of Frontier with 8 GPUs.

For the PMB model:
-------------------
Changes to the force calculation reduces:
- force calculation time from 313.90 s to 294.73 s (~6.1% reduction)
- communication time from 5.12 s to 3.82 s (~25.4% reduction)
- total time from 339.62 s to 316.45 s (~6.8% reduction)

For the LPS model:
------------------
(a) Changes to the weighted volume calculation reduces communication time from 383.48 s to 371.92 s (~3% reduction)
(b) Changes to the dilatation calculation in addition to (a) reduces communication time from 383.48 s to 350.14 s (~8.7 % reduction)
(c) Changes to the force calculation in addition to (a) and (b) reduces:
- force calculation time from 394.40 s to 374.19 s (~5.1% reduction)
- communication time from 383.48 s to 350.57 s (~8.6% reduction)
- total time from 801.35 s to 769.50 s (~4% reduction)