@streeve 2023-08-16
Also worth checking how much of a slowdown this is for 1 CPU node