@pabloseleson 2023-08-17
Performance change on 1 CPU-node of Frontier (56 CPUs) after the added force modifications:

For the PMB model:
-------------------
- force calculation time : 1280.99 s  -> 1258.08 s 
- communication time: 6.42 s -> 15.70 s 
- total time: 1344.85 s -> 1347.19 s

For the LPS model:
-------------------
- force calculation time : 3649.69 s  -> 3486.72 s 
- communication time: 55.31 s -> 68.27 s 
- total time: 3795.51 s -> 3688.12 s